requirements.txt
Flask
flask_cors
mysql-connector-python
opencv-python
face_recognition
numpy
Werkzeug
